Extract shared runtime and runner startup helpers
