+++ crates/core/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "supersonic-core"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+name = "supersonic_core"
+path = "src/lib.rs"
+
+[dependencies]
+gpu-hal = { path = "../gpu-hal" }
