+++ crates/core/src/lib.rs
@@ -1,0 +1,5 @@
+//! Shared model/backend metadata for SuperSonic frontends and runtimes.
+
+pub mod backend;
+pub mod capabilities;
+pub mod registry;
