+++ crates/runner/src/oracle.rs
@@ -1,3 +1,5 @@
+#![allow(dead_code)]
+
 use std::env;
 use std::path::Path;
 use std::process::Command;
