+++ crates/runner/tests/metal_qwen_smoke.rs
@@ -1,4 +1,6 @@
+#[cfg(target_os = "macos")]
 use std::path::PathBuf;
+#[cfg(target_os = "macos")]
 use std::process::Command;
 
 #[cfg(target_os = "macos")]
