+++ crates/runtime/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "supersonic-runtime"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+name = "supersonic_runtime"
+path = "src/lib.rs"
+
+[dependencies]
+supersonic-core = { path = "../core" }
+runner = { path = "../runner" }
+gpu-hal = { path = "../gpu-hal" }
+kernel-ffi = { path = "../kernel-ffi" }
+model-store = { path = "../model-store" }
+qwen35 = { path = "../qwen35" }
+gemma4 = { path = "../gemma4" }
+
+anyhow = "1"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+tokenizers = "0.22"
+tokio = { version = "1", features = ["rt-multi-thread", "macros", "sync"] }
+tracing = "0.1"
+rand = "0.8"
+rand_chacha = "0.3"
+minijinja = { version = "2", features = ["loop_controls"] }
+minijinja-contrib = { version = "2", features = ["pycompat"] }
