+++ crates/runtime/src/chat_template.rs
